Do not paint fully transparent nodes
With the current renderer we paint everything, including nodes that are
fully transparent.  Granted since the thing we paint is also transparent
it doesn't effect the output other than wasting cycles.

A limit of the current approach is that we can not easily drop branches
of the tree to be rendered when a parent node is fully transparent.

This also fixes the opacity node implementation in general to respect any
previously set opacity.

Change-Id: I993fcd4db5b714a3e8caf60135add0bb660c0b80
Reviewed-by: Lars Knoll <lars.knoll@digia.com>