Fix text rendering
Since we don't save/restore the painter state for each node, we have to be
careful that settings don't "leak" across paint calls. For example a rectangle
node with a gradient will leave the painter with a brush set.  A sub-sequent
child text node will be rendered incorrectly. This patch corrects that by
setting the brush explicitly to the empty brush.

This fixes also the examples/quick/views rendering.

Change-Id: Ia1415b5fdf77a560b02719917d36298e808418f1
Reviewed-by: Louai Al-Khanji <louai.al-khanji@digia.com>
Reviewed-by: Lars Knoll <lars.knoll@digia.com>