Correct painting of RectangleNode
The previous behavior was very wrong compared to how QtQuick normally
paints Rectangle elements.  The border.width property does not effect
the geometry of the Rectangle, where as QPainter::drawRect draws a
border 0.5 the border.width around the edge of the rectangle.  So now
there is special logic to draw the each case correctly using fillRect as
much as possible.

Change-Id: I4ea2be2226ff3b1a11dc5000522df503c6f43227
Reviewed-by: Lars Knoll <lars.knoll@digia.com>