Fix not having a context when making a QQuickWidget visible again
If we hit the size-is-zero path, the condition that got recently introduced
(if fakeHidden && d->context) is not really sufficient. It could be that
d->context is valid but the scenegraph is invalidated (i.e.
offscreenWindow->openglContext() is null).

This case has to be handled the same way.

Therefore, the function is now simplified to perform polish & sync
in one place.

Task-number: QTBUG-40794
Change-Id: Ia54f25fbdc199a0b88501a2d89edb5cee5411972
Reviewed-by: Paul Olav Tvete <paul.tvete@digia.com>