Add new property "designersupported" to qmldir
This patch adds a property called "designersupported" to qmldir.

This allows the Qt Quick Designer to only load plugins that
have the line ""designersupported"" in their qmldir file.

So the designer can load sub components without risking to load
plugins that have never been tested in the designer and that might crash.

The check for "designersupported"" is activated by using
QQmlImports::setDesignerSupportRequired().

Change-Id: I4bf07cc163faa47996eacb1365a7961c51c51060
Reviewed-by: Thomas Hartmann <Thomas.Hartmann@digia.com>
Reviewed-by: Tim Jenssen <tim.jenssen@digia.com>