Provide a default SurfaceFormat for our Context
Since we are not using OpenGL the content of the SurfaceFormat does not
matter to us.  Instead it gives us a way to expose the fact that we are
rendering to a raster surface.  Using the OpenGLInfo API from QML allows
you to check the properties of the SurfaceFormat and see that the
QSurfaceFormat::RenderableType is not OpenGL or OpenGLES.

Change-Id: Iac087dd89115f3455b4fb1941385352cadd2a364
Reviewed-by: Lars Knoll <lars.knoll@digia.com>