Support rendering Gradients when stops are not between 0 and 1
QGradient does not allow stops positions that are outside the bounds of
0 and 1. Gradients created from QML however can have stops outside of
these bounds and so we must normalize these stops before giving them to
QPainter to render.  This change creates a new stop at 0 or 1 if a stop
goes below or above the bounds respectively, and interpolates a new
color value for the stop based on the original position.

Change-Id: I10cc41d923e88a9b41bfd27de0326cea25464add
Reviewed-by: Simon Hausmann <simon.hausmann@digia.com>