Made emission of AnimatedSprite currentFrameChanges consistent.
Both the order was wrong and sometimes the currentFrame property
did not correspond to the 'arg' argument that was carried by the signal.
The order was sometimes like this:

 [window got exposed, m_running == false]
currentFrame: 1, arg: 1		<== Unexpected, its not running
 [start() got called, m_running == true]
currentFrame: 1, arg: 0		<== Unexpected, they are inequal
currentFrame: 2, arg: 2		<== Not sure what to expect here

The root of the problem was that we would enter prepareNextFrame on
startup (due to QQuickWindow::exposeEvent) even if we were in the
stopped state. Getting rid of the first invalid emission solved
the issue.

Change-Id: Iba002ee91d4d1acafcf4fd0911c28c93a4bc6fd5
Reviewed-by: Alan Alpert <aalpert@blackberry.com>