Improve rendering quality of Rectangles
The existing implementation has both inaccurate drawing of
borders as well as many alignment issues.  Much of these
alignment issues are due to our use of floating point sizes
and the integer based nature of blitter hardware.  So now
for the Rectangle node we force alignment via the use of
whole numbers.  This should be a better compromise of
speed vs accuracy than previously.

Change-Id: I9d8e9d1b3347a99bf0c2ae65ec66c935d8017817
Reviewed-by: Simon Hausmann <simon.hausmann@digia.com>