Improve quality of RectangleNodes when rotated
When there is a rotation transformation on a RectangleNode it is not
possible to take advantage of accelerated blitter paths, but we still
need to do custom painting to recreate the behavior of a Rectangle
component in Qt Quick.  Now when there is a rotaiton transformation we
will fall back to using QPainter::drawRect as long as there is no border
or radius.  If there is a border or radius value set though we will
instead first render to a QPixmap before painting as the standard
behavior of QPainter::drawRect and QPainter::drawRoundedRect does not
match the expected behavior of a QQuick Rectangle item. This fixes the
poor quality of rotated Rectangles at the cost of speed.

The Rotation of any item should still be discouraged for performance
reasons, but at least now Rectangles itmes should look correct when
used.

Change-Id: I3919ad732c9b4354a95e28e8dc09fcefd53f604e
Reviewed-by: Eirik Aavitsland <eirik.aavitsland@digia.com>