Fix clip state tracking in Context2D
A recent change (f45fe58ad2aa741c90b756643da75f1a6bc2fdf6)
introduced the concept of clipping being disabled for a context,
so that it can start with no clip path set.  However, this state
is not properly tracked, so an attempt to restore a context state
which has no clip path does not work properly.  This change adds
this information to the Clip command, so that it can be properly
restored.

This patch also re-enables a test case which was supposed to
check this behavior, but had been disabled.

Task-number: QTBUG-40312
Change-Id: I3fd5626ecfcc1298a81931828cbb590290098a92
Reviewed-by: Ulf Hermann <ulf.hermann@digia.com>
Reviewed-by: Gunnar Sletta <gunnar@sletta.org>